Refactor file operation hierarchy to unify under AsyncOperation
Replaced the `_Operation` base class with `AsyncOperation` to simplify and consolidate asynchronous operation handling. Cleaned up references, adjusted imports, and updated type annotations across related modules for consistency. Removed the unused `_Operation` class to reduce redundancy.
