+++ dev_kit_mcp_server/tools/core/__init__.py
@@ -1,9 +1,8 @@
 """core repo operations."""
 
-from dev_kit_mcp_server.tools.core.file_base import _Operation
-from dev_kit_mcp_server.tools.core.ops_async import AsyncOperation
+from dev_kit_mcp_server.tools.core.base_op import AsyncOperation
 
 __all__ = [
     "AsyncOperation",
-    "_Operation",
+    # "_Operation",
 ]
